requirements.txt
# for data generation
python-dotenv>=0.19.0
anthropic==0.45.0

# ML dependencies
torch>=2.1.0
transformers==4.48.1
peft==0.7.1
bitsandbytes==0.43.2
accelerate>=0.26.0
datasets>=3.0.0

# Development dependencies
flake8>=6.1.0
py-cpuinfo==9.0.0
psutil==5.9.5

# for test.py
IPython
ipywidgets
